feat: move bridge into wallet and add debridge option
