+++ apps/marginfi-v2-ui/src/components/common/Bridge/index.tsx
@@ -1,0 +1,1 @@
+export * from "./Bridge";
