+++ apps/marginfi-v2-ui/src/global.d.ts
@@ -1,14 +1,17 @@
 interface Window {
   MayanSwap: {
-    init: (id: string, config: import('./types').MayanWidgetConfigType) => void;
-    updateConfig: (newConfig: import('./types').MayanWidgetConfigType | import('./types').MayanWidgetSolanaConfigType) => void;
-    setSwapInitiateListener(callback: (swap: import('./types').MayanSwapInfo) => void): void;
+    init: (id: string, config: import("./types").MayanWidgetConfigType) => void;
+    updateConfig: (
+      newConfig: import("./types").MayanWidgetConfigType | import("./types").MayanWidgetSolanaConfigType
+    ) => void;
+    setSwapInitiateListener(callback: (swap: import("./types").MayanSwapInfo) => void): void;
     removeSwapInitiateListener(): void;
-    setSwapCompleteListener(callback: (swap: import('./types').MayanSwapInfo) => void): void;
+    setSwapCompleteListener(callback: (swap: import("./types").MayanSwapInfo) => void): void;
     removeSwapCompleteListener(): void;
-    setSwapRefundListener(callback: (swap: import('./types').MayanSwapInfo) => void): void;
+    setSwapRefundListener(callback: (swap: import("./types").MayanSwapInfo) => void): void;
     removeSwapRefundListener(): void;
   };
   Jupiter: any;
   MoonPayWebSdk: any;
+  deBridge: any;
 }
